+++ lib/mutant/mutator/node/define.rb
@@ -4,8 +4,25 @@ module Mutant
       class Define < self
 
         handle(Rubinius::AST::Define)
-        handle(Rubinius::AST::DefineSingleton)
         handle(Rubinius::AST::DefineSingletonScope)
+
+      private
+
+        # Emit mutations
+        #
+        # @return [undefined]
+        #
+        # @api private
+        #
+        def dispatch
+          emit_attribute_mutations(:body)
+          emit_attribute_mutations(:arguments)
+        end
+      end
+
+      class DefineSingleton < self
+
+        handle(Rubinius::AST::DefineSingleton)
 
       private