+++ spec/unit/mutant/mutator/node/define/mutation_spec.rb
@@ -23,6 +23,35 @@ describe Mutant::Mutator, 'define' do
     it_should_behave_like 'a mutator'
   end
 
+  context 'with arguments' do
+    let(:source) { 'def foo(a, b); nil; end' }
+
+    before do
+      Mutant::Random.stub(:hex_string => 'random')
+    end
+
+    let(:mutations) do
+      mutations = []
+
+      # Deletion of each argument
+      mutations << 'def foo(a); nil; end'
+      mutations << 'def foo(b); nil; end'
+
+      # Deletion of all arguments
+      mutations << 'def foo; nil; end'
+
+      # Rename each argument
+      mutations << 'def foo(srandom, b); nil; end'
+      mutations << 'def foo(a, srandom); nil; end'
+
+      # Mutation of body
+      mutations << 'def foo(a, b); Object.new; end'
+    end
+
+
+    it_should_behave_like 'a mutator'
+  end
+
   context 'define on singleton' do
     let(:source) { 'def self.foo; self.bar; self.baz; end' }
 
@@ -44,4 +73,32 @@ describe Mutant::Mutator, 'define' do
     it_should_behave_like 'a mutator'
   end
 
+  context 'define on singleton with argument' do
+
+    before do
+      Mutant::Random.stub(:hex_string => 'random')
+    end
+
+    let(:source) { 'def self.foo(a, b); nil; end' }
+
+    let(:mutations) do
+      mutations = []
+
+      # Deletion of each argument
+      mutations << 'def self.foo(a); nil; end'
+      mutations << 'def self.foo(b); nil; end'
+
+      # Deletion of all arguments
+      mutations << 'def self.foo; nil; end'
+
+      # Rename each argument
+      mutations << 'def self.foo(srandom, b); nil; end'
+      mutations << 'def self.foo(a, srandom); nil; end'
+
+      # Mutation of body
+      mutations << 'def self.foo(a, b); Object.new; end'
+    end
+
+    it_should_behave_like 'a mutator'
+  end
 end